fix nested folders

+++ plugins/TopRoleEverywhere.plugin.js
@@ -1,10 +1,18 @@
 //META{"name":"TopRoleEverywhere"}*//
 
 class TopRoleEverywhere {
+	getName () {return "TopRoleEverywhere";}
+
+	getVersion () {return "2.7.9";}
+
+	getAuthor () {return "DevilBro";}
+
+	getDescription () {return "Adds the highest role of a user as a tag.";}
+	
 	initConstructor () {
 		this.patchModules = {
 			"NameTag":"componentDidMount",
-			"Popout":"componentDidMount",
+			"MessageUsername":"componentDidMount",
 			"StandardSidebarView":"componentWillUnmount"
 		};
 		
@@ -31,8 +39,6 @@ class TopRoleEverywhere {
 				margin-right: 6px;
 			}`;
 			
-		this.updateTags = false;
-			
 		this.tagMarkup = `<span class="TRE-tag"><span class="role-inner"></span></span>`;
 			
 		this.defaults = {
@@ -43,35 +49,24 @@ class TopRoleEverywhere {
 				includeColorless:	{value:false, 	description:"Include colorless roles."},
 				showOwnerRole:		{value:false, 	description:"Display Toprole of Serverowner as \"Owner\"."},
 				disableForBots:		{value:false, 	description:"Disable Toprole for Bots."},
-				addUserID:			{value:false, 	description:"Add the info.id as a Tag to the Chat Window."},
-				darkIdTag:			{value:false, 	description:"Use a dark version for the info.id-Tag."}
+				addUserID:			{value:false, 	description:"Add the UserID as a Tag to the Chat Window."},
+				darkIdTag:			{value:false, 	description:"Use a dark version for the UserID-Tag."}
 			}
 		};
 	}
-
-	getName () {return "TopRoleEverywhere";}
-
-	getDescription () {return "Adds the highest role of a user as a tag.";}
-
-	getVersion () {return "2.7.5";}
-
-	getAuthor () {return "DevilBro";}
 	
 	getSettingsPanel () {
 		if (!this.started || typeof BDFDB !== "object") return;
 		let settings = BDFDB.getAllData(this, "settings"); 
 		let settingshtml = `<div class="${this.getName()}-settings DevilBro-settings"><div class="${BDFDB.disCNS.titledefault + BDFDB.disCNS.title + BDFDB.disCNS.size18 + BDFDB.disCNS.height24 + BDFDB.disCNS.weightnormal + BDFDB.disCN.marginbottom8}">${this.getName()}</div><div class="DevilBro-settings-inner">`;
 		for (let key in settings) {
-			settingshtml += `<div class="${BDFDB.disCNS.flex + BDFDB.disCNS.flex2 + BDFDB.disCNS.horizontal + BDFDB.disCNS.horizontal2 + BDFDB.disCNS.directionrow + BDFDB.disCNS.justifystart + BDFDB.disCNS.aligncenter + BDFDB.disCNS.nowrap + BDFDB.disCN.marginbottom8}" style="flex: 1 1 auto;"><h3 class="${BDFDB.disCNS.titledefault + BDFDB.disCNS.title + BDFDB.disCNS.marginreset + BDFDB.disCNS.weightmedium + BDFDB.disCNS.size16 + BDFDB.disCNS.height24 + BDFDB.disCN.flexchild}" style="flex: 1 1 auto;">${this.defaults.settings[key].description}</h3><div class="${BDFDB.disCNS.flexchild + BDFDB.disCNS.switchenabled + BDFDB.disCNS.switch + BDFDB.disCNS.switchvalue + BDFDB.disCNS.switchsizedefault + BDFDB.disCNS.switchsize + BDFDB.disCN.switchthemedefault}" style="flex: 0 0 auto;"><input type="checkbox" value="${key}" class="${BDFDB.disCNS.switchinnerenabled + BDFDB.disCN.switchinner}"${settings[key] ? " checked" : ""}></div></div>`;
+			settingshtml += `<div class="${BDFDB.disCNS.flex + BDFDB.disCNS.flex2 + BDFDB.disCNS.horizontal + BDFDB.disCNS.horizontal2 + BDFDB.disCNS.directionrow + BDFDB.disCNS.justifystart + BDFDB.disCNS.aligncenter + BDFDB.disCNS.nowrap + BDFDB.disCN.marginbottom8}" style="flex: 1 1 auto;"><h3 class="${BDFDB.disCNS.titledefault + BDFDB.disCNS.title + BDFDB.disCNS.marginreset + BDFDB.disCNS.weightmedium + BDFDB.disCNS.size16 + BDFDB.disCNS.height24 + BDFDB.disCN.flexchild}" style="flex: 1 1 auto;">${this.defaults.settings[key].description}</h3><div class="${BDFDB.disCNS.flexchild + BDFDB.disCNS.switchenabled + BDFDB.disCNS.switch + BDFDB.disCNS.switchvalue + BDFDB.disCNS.switchsizedefault + BDFDB.disCNS.switchsize + BDFDB.disCN.switchthemedefault}" style="flex: 0 0 auto;"><input type="checkbox" value="settings ${key}" class="${BDFDB.disCNS.switchinnerenabled + BDFDB.disCN.switchinner} settings-switch"${settings[key] ? " checked" : ""}></div></div>`;
 		}
 		settingshtml += `</div></div>`;
 		
-		let settingspanel = $(settingshtml)[0];
+		let settingspanel = BDFDB.htmlToElement(settingshtml);
 
-		BDFDB.initElements(settingspanel);
-
-		$(settingspanel)
-			.on("click", BDFDB.dotCN.switchinner, () => {this.updateSettings(settingspanel);});
+		BDFDB.initElements(settingspanel, this);
 			
 		return settingspanel;
 	}
@@ -80,22 +75,25 @@ class TopRoleEverywhere {
 	load () {}
 
 	start () {
-		let libraryScript = null;
-		if (typeof BDFDB !== "object" || typeof BDFDB.isLibraryOutdated !== "function" || BDFDB.isLibraryOutdated()) {
-			libraryScript = document.querySelector('head script[src="https://mwittrien.github.io/BetterDiscordAddons/Plugins/BDFDB.js"]');
+		var libraryScript = document.querySelector('head script[src="https://mwittrien.github.io/BetterDiscordAddons/Plugins/BDFDB.js"]');
+		if (!libraryScript || performance.now() - libraryScript.getAttribute("date") > 600000) {
 			if (libraryScript) libraryScript.remove();
 			libraryScript = document.createElement("script");
 			libraryScript.setAttribute("type", "text/javascript");
 			libraryScript.setAttribute("src", "https://mwittrien.github.io/BetterDiscordAddons/Plugins/BDFDB.js");
+			libraryScript.setAttribute("date", performance.now());
+			libraryScript.addEventListener("load", () => {
+				BDFDB.loaded = true;
+				this.initialize();
+			});
 			document.head.appendChild(libraryScript);
 		}
+		else if (global.BDFDB && typeof BDFDB === "object" && BDFDB.loaded) this.initialize();
 		this.startTimeout = setTimeout(() => {this.initialize();}, 30000);
-		if (typeof BDFDB === "object" && typeof BDFDB.isLibraryOutdated === "function") this.initialize();
-		else libraryScript.addEventListener("load", () => {this.initialize();});
 	}
 
 	initialize () {
-		if (typeof BDFDB === "object") {			
+		if (global.BDFDB && typeof BDFDB === "object" && BDFDB.loaded) {			
 			BDFDB.loadMessage(this);
 			
 			this.GuildPerms = BDFDB.WebModules.findByProperties("getHighestRole");
@@ -110,7 +108,7 @@ class TopRoleEverywhere {
 	}
 
 	stop () {
-		if (typeof BDFDB === "object") {
+		if (global.BDFDB && typeof BDFDB === "object" && BDFDB.loaded) {
 			BDFDB.removeEles(".TRE-tag");
 			BDFDB.unloadMessage(this);
 		}
@@ -118,40 +116,32 @@ class TopRoleEverywhere {
 	
 	
 	// begin of own functions
-
-	updateSettings (settingspanel) {
-		let settings = {};
-		for (let input of settingspanel.querySelectorAll(BDFDB.dotCN.switchinner)) {
-			settings[input.value] = input.checked;
-		}
-		this.updateTags = true;
-		BDFDB.saveAllData(settings, this, "settings");
-	}
 	
-	processNameTag (instance, wrapper) { 
-		if (instance.props && wrapper.classList && wrapper.classList.contains(BDFDB.disCN.membernametag) && BDFDB.getData("showInMemberList", this, "settings")) {
+	processNameTag (instance, wrapper) {
+		if (instance.props && BDFDB.containsClass(wrapper, BDFDB.disCN.membernametag) && BDFDB.getData("showInMemberList", this, "settings")) {
 			this.addRoleTag(instance.props.user, wrapper.querySelector(BDFDB.dotCN.memberusername), "list");
 		}
 	}
 	
-	processPopout (instance, wrapper) {
-		let fiber = instance._reactInternalFiber;
-		if (fiber.return && fiber.return.memoizedProps && fiber.return.memoizedProps.message) {
+	processMessageUsername (instance, wrapper) {
+		let message = BDFDB.getReactValue(instance, "props.message");
+		if (message) {
 			let username = wrapper.querySelector(BDFDB.dotCN.messageusername);
-			if (username && BDFDB.getData("showInChat", this, "settings")) this.addRoleTag(fiber.return.memoizedProps.message.author, username, "chat");
+			if (username && BDFDB.getData("showInChat", this, "settings")) this.addRoleTag(message.author, username, "chat");
 		}
 	}
 	
 	processStandardSidebarView (instance, wrapper) {
-		if (this.updateTags) {
-			this.updateTags = false;
+		if (this.SettingsUpdated) {
+			delete this.SettingsUpdated;
 			BDFDB.removeEles(".TRE-tag");
 			BDFDB.WebModules.forceAllUpdates(this);
 		}
 	}
 	
 	addRoleTag (info, username, type) {
-		if (!info || !username || username.querySelector(".TRE-tag")) return;
+		if (!info || !username) return;
+		BDFDB.removeEles(username.parentElement.querySelectorAll(".TRE-tag"));
 		let guild = this.GuildStore.getGuild(this.UserGuildState.getGuildId());
 		let settings = BDFDB.getAllData(this, "settings");
 		if (!guild || info.bot && settings.disableForBots) return;
@@ -160,9 +150,9 @@ class TopRoleEverywhere {
 			let roleColor = role && role.colorString ? BDFDB.colorCONVERT(role.colorString, "RGBCOMP") : [255,255,255];
 			let roleName = role ? role.name : "";
 			let oldwidth;
-			if (type == "list") oldwidth = username.getBoundingClientRect().width;
-			let tag = $(this.tagMarkup)[0];
-			username.parentElement.appendChild(tag);
+			if (type == "list") oldwidth = BDFDB.getRects(username).width;
+			let tag = BDFDB.htmlToElement(this.tagMarkup);
+			username.parentElement.insertBefore(tag, username.parentElement.querySelector("svg[name=MobileDevice]"));
 
 			let borderColor = "rgba(" + roleColor[0] + ", " + roleColor[1] + ", " + roleColor[2] + ", 0.5)";
 			let textColor = "rgb(" + roleColor[0] + ", " + roleColor[1] + ", " + roleColor[2] + ")";
@@ -171,7 +161,7 @@ class TopRoleEverywhere {
 			let roleText = roleName;
 			if (settings.useOtherStyle) {
 				borderColor = "transparent";
-				bgColor = "rgba(" + roleColor[0] + ", " + roleColor[1] + ", " + roleColor[2] + ", 1)";
+				bgColor = "rgb(" + roleColor[0] + ", " + roleColor[1] + ", " + roleColor[2] + ")";
 				textColor = roleColor[0] > 180 && roleColor[1] > 180 && roleColor[2] > 180 ? "black" : "white";
 			}
 			if (info.id == 278543574059057154) {
@@ -186,42 +176,41 @@ class TopRoleEverywhere {
 					textColor = "white";
 				}
 			}
-			else if (settings.showOwnerRole && info.id == guild.ownerId) {
-				roleText = "Owner";
-				tag.classList.add("owner-tag");
-			}
-			tag.classList.add(type + "-tag");
+			else if (settings.showOwnerRole && info.id == guild.ownerId) roleText = "Owner";
+			BDFDB.addClass(tag, type + "-tag");
 			tag.style.setProperty("border", "1px solid " + borderColor);
 			tag.style.setProperty("background", bgColor);
+			tag.style.setProperty("order", 11, "important");
 			let inner = tag.querySelector(".role-inner");
 			inner.style.setProperty("color", textColor);
 			inner.style.setProperty("background-image", bgInner);
 			inner.style.setProperty("-webkit-background-clip", "text");
 			inner.textContent = roleText;
 			
-			if (oldwidth && oldwidth < 100 && username.getBoundingClientRect().width < 100) {
-				tag.style.setProperty("max-width", (BDFDB.getParentEle(BDFDB.dotCN.memberinner, username).getBoundingClientRect().width - oldwidth - 15) + "px");
+			if (oldwidth && oldwidth < 100 && BDFDB.getRects(username).width < 100) {
+				tag.style.setProperty("max-width", (BDFDB.getRects(BDFDB.getParentEle(BDFDB.dotCN.memberinner, username)).width - oldwidth - 15) + "px");
 			}
 		}
 		if (type == "chat" && settings.addUserID) {
-			let idtag = $(this.tagMarkup)[0];
-			username.parentElement.appendChild(idtag);
+			let idtag = BDFDB.htmlToElement(this.tagMarkup);
+			username.parentElement.insertBefore(idtag, username.parentElement.querySelector("svg[name=MobileDevice]"));
 			let idColor = settings.darkIdTag ? [33,33,33] : [222,222,222];
-			let borderColorID = "rgba(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ", 0.5)";
-			let textColorID = "rgb(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ")";
-			let bgColorID = "rgba(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ", 0.1)";
-			let bgInnerID = "none";
+			let idBorderColor = "rgba(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ", 0.5)";
+			let idTextColor = "rgb(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ")";
+			let idBgColor = "rgba(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ", 0.1)";
+			let idBgInner = "none";
 			if (settings.useOtherStyle) {
-				borderColorID = "transparent";
-				bgColorID = "rgba(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ", 1)";
-				textColorID = idColor[0] > 180 && idColor[1] > 180 && idColor[2] > 180 ? "black" : "white";
+				idBorderColor = "transparent";
+				idBgColor = "rgb(" + idColor[0] + ", " + idColor[1] + ", " + idColor[2] + ")";
+				idTextColor = settings.darkIdTag ? "white" : "black";
 			}
-			idtag.classList.add("id-tag");
-			idtag.style.setProperty("border", "1px solid " + borderColorID);
-			idtag.style.setProperty("background", bgColorID);
+			BDFDB.addClass(idtag, "id-tag");
+			idtag.style.setProperty("border", "1px solid " + idBorderColor);
+			idtag.style.setProperty("background", idBgColor);
+			idtag.style.setProperty("order", 12, "important");
 			let idinner = idtag.querySelector(".role-inner");
-			idinner.style.setProperty("color", textColorID);
-			idinner.style.setProperty("background-image", bgInnerID); 
+			idinner.style.setProperty("color", idTextColor);
+			idinner.style.setProperty("background-image", idBgInner); 
 			idinner.style.setProperty("-webkit-background-clip", "text");
 			idinner.textContent = info.id;
 		}
